Move transform to utils, comments
