+++ modeltrans/utils.py
@@ -1,3 +1,5 @@
+from django.contrib.postgres.fields.jsonb import KeyTransform
+from django.db.models.lookups import Transform
 from django.utils.translation import get_language as _get_language
 
 from .conf import get_available_languages, get_default_language
@@ -26,3 +28,39 @@ def build_localized_fieldname(field_name, lang):
         # current naming scheme as Django foreign keys also add "id" suffix.
         lang = "ind"
     return "{}_{}".format(field_name, lang.replace("-", "_"))
+
+
+class FallbackTransform(Transform):
+    """
+    Custom version of KeyTextTransform to use a database field as part of the key.
+
+    For example: with default_language="nl", calling
+    `FallbackTransformb("title_", F("fallback_language"), "i18n")` becomes in SQL"
+        `"i18n"->>('title_' || "fallback_language")`
+    """
+
+    def __init__(self, field_prefix, language_expression, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.field_prefix = field_prefix
+        self.language_expression = language_expression
+
+    def preprocess_lhs(self, compiler, connection, lhs_only=False):
+        if not lhs_only:
+            key_transforms = [self.field_prefix]
+        previous = self.lhs
+        while isinstance(previous, KeyTransform):
+            if not lhs_only:
+                key_transforms.insert(0, previous.key_name)
+            previous = previous.lhs
+        lhs, params = compiler.compile(previous)
+        return (lhs, params, key_transforms) if not lhs_only else (lhs, params)
+
+    def as_postgresql(self, compiler, connection):
+        lhs, params, key_transforms = self.preprocess_lhs(compiler, connection)
+        params.extend([self.field_prefix])
+
+        rhs = self.language_expression.resolve_expression(compiler.query)
+        rhs_sql, rhs_params = compiler.compile(rhs)
+        params.extend(rhs_params)
+
+        return "(%s ->> (%%s || %s ))" % (lhs, rhs_sql), (params)
